+++ j10/ex09/calculator.c
@@ -6,31 +6,31 @@
 /*   By: alcaroff <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/24 14:04:52 by alcaroff          #+#    #+#             */
-/*   Updated: 2017/07/25 11:39:16 by alcaroff         ###   ########.fr       */
+/*   Updated: 2017/07/26 09:47:48 by alcaroff         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_mul(int a, int b)
+long	ft_mul(long a, long b)
 {
 	return (a * b);
 }
 
-int		ft_div(int a, int b)
+long	ft_div(long a, long b)
 {
 	return (a / b);
 }
 
-int		ft_sub(int a, int b)
+long	ft_sub(long a, long b)
 {
 	return (a - b);
 }
 
-int		ft_add(int a, int b)
+long	ft_add(long a, long b)
 {
 	return (a + b);
 }
 
-int		ft_mod(int a, int b)
+long	ft_mod(long a, long b)
 {
 	return (a % b);
 }
